add signInWithEmail of firebase.auth function
