+++ src/controller/login-controller.js
@@ -1,12 +1,12 @@
-import loginUserEmail from '../model/register-model.js';
-import loginView from '../view/login-view.js';
+import { loginUserEmail } from '../model/login-model.js';
 
 export default () => {
-  loginView();
-  const btnLogin = document.querySelector('.btn-login');
-  btnLogin.addEventListener('click', () => {
-    const email = document.querySelector('#email-login').value;
-    const password = document.querySelector('#password-login').value;
-    loginUserEmail(email, password);
+  const btnLogin = document.querySelector('#btn-login');
+  btnLogin.addEventListener('click', (event) => {
+    event.preventDefault();
+    const email = document.querySelector('#email-login');
+    const password = document.querySelector('#password-login');
+    loginUserEmail(email.value, password.value);
+    
   });
 };
