+++ src/controller/router.js
@@ -1,5 +1,6 @@
 import { components } from '../view/index.js';
 import registerController from './register-controller.js';
+import loginController from './login-controller.js';
 
 export const changeView = (hash) => {
   const container = document.querySelector('#container');
@@ -8,6 +9,7 @@ export const changeView = (hash) => {
     case '':
     case '#/login':
       container.appendChild(components.login());
+      loginController();
       break;
     case '#/register':
       container.appendChild(components.register());
