+++ src/model/register-model.js
@@ -7,12 +7,3 @@ export const registerUserEmail = (email, password) => {
     console.log(errorMessage);
   });
 };
-export const loginUserEmail = (email, password) => {
-  firebase.auth().signInWithEmailAndPassword(email, password).catch((error) => {
-    // Handle Errors here.
-    const errorCode = error.code;
-    const errorMessage = error.message;
-    console.log(errorCode);
-    console.log(errorMessage);
-  });
-};
